--- a/queries.sql
+++ b/queries.sql
@@ -501,3 +501,7 @@ ORDER BY
     number_of_animals DESC
 LIMIT
     1;
+
+EXPLAIN ANALYZE SELECT COUNT(*) FROM visits where animal_id = 4;
+EXPLAIN ANALYZE SELECT * FROM visits where vet_id = 2;
+EXPLAIN ANALYZE SELECT * FROM owners where email = 'owner_18327@mail.com';