--- a/schema.sql
+++ b/schema.sql
@@ -56,3 +56,8 @@ CREATE TABLE visits (
     FOREIGN KEY (animal_id) REFERENCES animals(id),
     FOREIGN KEY (vet_id) REFERENCES vets(id)
 );
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+
+CREATE INDEX index_animalID ON visits(animal_id);
+CREATE INDEX index_vetID ON visits(vet_id);
+CREATE INDEX indx_Email on owners(email);