In this Pull Request, we have added several optimizations to improve the execution time of the provided queries.

After we ran the scripts that seed data, the **visits** and **owners** tables had 21,565,700 and 15,000,006 rows respectively, and the database grew to **2,778MB**.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/e7ca087f-06cf-4fde-9e6b-8f98575218d1)

## Query 1 
###  ```SELECT COUNT(*) FROM visits where animal_id = 4;```

The first run of our query, without any optimizations, took **714ms**. 

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/bb67fc40-bcfe-4fc3-97f5-8980e81b7ff5)

After this, we used the following command to index the **visits** table on the **animal_id** column.

```CREATE INDEX index_animalID ON visits(animal_id);```

We can see the index was created in this screenshot:

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/bb4ded9b-26c3-4e9c-8c74-66dec65380d8)

Our query running time improved to **215ms**. This is a reduction of **69%**.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/b0a3f575-d0c6-4f00-9c38-e0030ec50eb1)


## Query 2
### ```SELECT * FROM visits where vet_id = 2;```

The first run of our query, without any optimizations, took **1,549ms**. 

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/a3e96958-6dd6-4e9d-addf-37832a568d6f)

After this, we used the following command to index the **visits** table on the **vet_id** column.

```CREATE INDEX index_vetID ON visits(vet_id);```

We can see the index was created in this screenshot:

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/bb4ded9b-26c3-4e9c-8c74-66dec65380d8)

However, even after indexing, this query runtime did not improve significantly.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/7f163ae5-c2fe-49c5-912c-ba571ce6b0c9)

Hence, we tried to use **table clustering** which physically reorders the rows based on the values in one or more columns, which is known as the _clustering key_. The primary goal of clustering is to group together rows with similar values in the clustering key column(s) on disk, making it more efficient to read and retrieve these rows when performing range scans or filtering based on the clustering key.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/3967f843-e07c-415c-984b-df3bbcd26a1f)

Once again, the query time did not improve substantially. Furthermore, running the query repeatedly gave us different results. The following is the best result we managed to get.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/3750e5b1-570a-4db7-9891-a63b9fbfbc35)

The reasons the running time of this query didn't improve significantly may be:
1. There are only four vets. Hence, indexing by the vet_id was not effective. 
2. The query had to fetch 5,391,429 records, and such an operation inherently takes a long time to complete.
3. The screenshot of EXPLAIN ANALYZE shows that PostgreSQL used sequential search even though we defined an index. This may be because the indexed column has low cardinality (a small number of distinct values relative to the total number of rows in the table, that is 4 versus  5,391,429). 

The following screenshot shows that there are a total of 21,565,700 records in the **visits** table which have only 4 distinct values for the **vet_id** column. Further, it shows the given query returns 5,391,429 records.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/e6473349-9651-47bd-a036-b88be263fdc4)


## Query 3
### ```SELECT * FROM owners where email = 'owner_18327@mail.com';```

The first run of our query, without any optimizations, took **663ms**. 

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/7b885bd6-2f88-41e8-9fae-d1629e2bc404)

After this, we used the following command to index the **owners** table on the **email** column.

```CREATE INDEX indx_Email on owners(email);```

We can see the index was created in this screenshot:

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/a8856bed-1e00-4ad1-9e4f-6e26a73848c4)

Our query running time improved to **3ms**. This is a reduction of **99.5%**.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/24445cba-f6fc-43eb-823e-48b2aa2fa401)

This performance gain was surprising. This may be because there are only 6 records with the specified email, out of a total of 15,000,006 records that have 2,500,000 unique email addresses. Hence, the index was highly effective.

![image](https://github.com/henokkhm/microverse-vet-clinic/assets/110997325/58c7614a-c187-4cb8-a457-b7e03f385343)